requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
flask-cors==4.0.0
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
playwright
